backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "使用swmm-api处理swmm文件的后端程序"
authors = ["tangkaiwen00@gmail.com"]
package-mode = false # 关闭包模式,只是用作依赖安装

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
swmm-api = "^0.4.64"
pyproj = "^3.7.1"
pyswmm = "^2.0.1"
python-dotenv = "^1.1.1"
langchain = {extras = ["openai"], version = "^0.3.27"}
langgraph = {extras = ["postgres"], version = "^0.6.4"}
langsmith = "^0.4.12"
psycopg = "^3.2.9"
psycopg-pool = "^3.2.6"
langgraph-checkpoint-postgres = "^2.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
